pyproject.toml
[tool.poetry]
name = "discord-pretty-help"
version = "1.2.1"
description = "And nicer looking interactive help menu for discord.py"
authors = ["StroupBSlayen <29642143+stroupbslayen@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/stroupbslayen/discord-pretty-help"
keywords=["discord", "discord.py", "discord bot"]
packages = [
    {include = "pretty_help"}
]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
"discord.py" = "^1.5.1"
black = {version = "^20.8b1", allow-prereleases = true}
python-dotenv = "^0.15.0"

[tool.poetry.scripts]
test = "tests.test_discord_pretty_help:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
